# refactor-baseline/openapi.json
{
  "components": {
    "schemas": {
...
      "AssociateIn": {
        "properties": {
          "case_id": {
            "title": "Case Id",
            "type": "string"
          }
        },
        "required": [
          "case_id"
        ],
        "title": "AssociateIn",
        "type": "object"
      },
      "Body_create_feature_api_features_post": {
        "properties": {
          "confluence_children": {
            "default": true,
            "title": "Confluence Children",
            "type": "boolean"
          },
          "confluence_url": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Confluence Url",
            "type": "array"
          },
          "figma_url": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Figma Url",
            "type": "array"
          },
          "files": {
            "items": {
              "format": "binary",
              "type": "string"
            },
            "title": "Files",
            "type": "array"
          },
          "focus": {
            "default": "",
            "title": "Focus",
            "type": "string"
          },
          "key": {
            "default": "",
            "title": "Key",
            "type": "string"
          },
          "match_key": {
            "default": "",
            "title": "Match Key",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "project_id": {
            "default": "",
            "title": "Project Id",
            "type": "string"
          },
          "text": {
            "default": "",
            "title": "Text",
            "type": "string"
          },
          "total": {
            "default": 16,
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "name"
        ],
        "title": "Body_create_feature_api_features_post",
        "type": "object"
      },
      "Body_new_version_api_features__fid__versions_post": {
        "properties": {
          "confluence_children": {
            "default": true,
            "title": "Confluence Children",
            "type": "boolean"
          },
          "confluence_url": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Confluence Url",
            "type": "array"
          },
          "figma_url": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Figma Url",
            "type": "array"
          },
          "files": {
            "items": {
              "format": "binary",
              "type": "string"
            },
            "title": "Files",
            "type": "array"
          },
          "focus": {
            "default": "",
            "title": "Focus",
            "type": "string"
          },
          "key": {
            "default": "",
            "title": "Key",
            "type": "string"
          },
          "name": {
            "default": "",
            "title": "Name",
            "type": "string"
          },
          "replace": {
            "default": "false",
            "title": "Replace",
            "type": "string"
          },
          "text": {
            "default": "",
            "title": "Text",
            "type": "string"
          },
          "total": {
            "default": 16,
            "title": "Total",
            "type": "integer"
          }
        },
        "title": "Body_new_version_api_features__fid__versions_post",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MatchKeyIn": {
        "properties": {
          "match_key": {
            "default": "",
            "title": "Match Key",
            "type": "string"
          }
        },
        "title": "MatchKeyIn",
        "type": "object"
      },
...
      "RenameIn": {
        "properties": {
          "key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Key"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "title": "RenameIn",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "wardenIQ \u2014 Test Intelligence Platform",
    "version": "0.1.0-beta"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/features": {
      "get": {
        "operationId": "list_features_api_features_get",
        "parameters": [
          {
            "in": "query",
            "name": "project_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Features"
      },
      "post": {
        "operationId": "create_feature_api_features_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_feature_api_features_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Feature"
      }
    },
    "/api/features/{fid}": {
      "delete": {
        "operationId": "delete_feature_api_features__fid__delete",
        "parameters": [
          {
            "in": "path",
            "name": "fid",
            "required": true,
            "schema": {
              "title": "Fid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Feature"
      },
      "get": {
        "operationId": "get_feature_api_features__fid__get",
        "parameters": [
          {
            "in": "path",
            "name": "fid",
            "required": true,
            "schema": {
              "title": "Fid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Feature"
      },
      "patch": {
        "operationId": "rename_feature_api_features__fid__patch",
        "parameters": [
          {
            "in": "path",
            "name": "fid",
            "required": true,
            "schema": {
              "title": "Fid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RenameIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Rename Feature"
      }
    },
    "/api/features/{fid}/associate": {
      "post": {
        "operationId": "associate_api_features__fid__associate_post",
        "parameters": [
          {
            "in": "path",
            "name": "fid",
            "required": true,
            "schema": {
              "title": "Fid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssociateIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Associate"
      }
    },
...
    "/api/features/{fid}/match-key": {
      "post": {
        "description": "Set/clear a feature's manual PR match tag (editor+, gated by auth_gateway).\nPRs whose title/body contain the bracketed tag (e.g. [HOLDS]) auto-map to this\nfeature on the next poll -- for projects without a linked Jira epic.",
        "operationId": "set_feature_match_key_api_features__fid__match_key_post",
        "parameters": [
          {
            "in": "path",
            "name": "fid",
            "required": true,
            "schema": {
              "title": "Fid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MatchKeyIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Feature Match Key"
      }
    },
...
    "/api/features/{fid}/regenerate": {
      "post": {
        "description": "Re-run generation on an existing feature with a different depth/focus\n(adds new cases; dedup keeps existing ones).",
        "operationId": "regenerate_api_features__fid__regenerate_post",
        "parameters": [
          {
            "in": "path",
            "name": "fid",
            "required": true,
            "schema": {
              "title": "Fid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Body",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Regenerate"
      }
    },
...
    "/api/features/{fid}/versions": {
      "post": {
        "operationId": "new_version_api_features__fid__versions_post",
        "parameters": [
          {
            "in": "path",
            "name": "fid",
            "required": true,
            "schema": {
              "title": "Fid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_new_version_api_features__fid__versions_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "New Version"
      }
    },
...
  }
}